bookingsRes formet updated

+++ src/config/db.ts
@@ -39,7 +39,7 @@ const initDB = async () => {
         vehicle_id INTEGER NOT NULL REFERENCES vehicles(id) ON DELETE CASCADE,
         rent_start_date DATE NOT NULL,
         rent_end_date DATE NOT NULL CHECK (rent_end_date > rent_start_date),
-        total_price NUMERIC(10,2) NOT NULL CHECK (total_price > 0),
+        total_price INTEGER NOT NULL CHECK (total_price > 0),
         status VARCHAR(20) NOT NULL CHECK (status IN ('active','cancelled','returned'))
         )
         `);
